+++ pkg/otlp/attributes/source_test.go
@@ -27,6 +27,7 @@ import (
 )
 
 const (
+	testLiteralHost            = "literal-host"
 	testHostID                 = "example-host-id"
 	testHostName               = "example-host-name"
 	testContainerID            = "example-container-id"
@@ -38,151 +39,6 @@ const (
 	testGCPIntegrationHostname = testHostName + "." + testCloudAccount
 )
 
-func TestSourceFromAttributes(t *testing.T) {
-	tests := []struct {
-		name       string
-		attrs      pcommon.Map
-		usePreview bool
-
-		ok  bool
-		src source.Source
-	}{
-		{
-			name: "custom hostname",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				AttributeDatadogHostname:            testCustomName,
-				AttributeK8sNodeName:                testNodeName,
-				conventions.AttributeK8SClusterName: testClusterName,
-				conventions.AttributeContainerID:    testContainerID,
-				conventions.AttributeHostID:         testHostID,
-				conventions.AttributeHostName:       testHostName,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testCustomName},
-		},
-		{
-			name: "container ID",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeContainerID: testContainerID,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testContainerID},
-		},
-		{
-			name: "container ID, preview",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeContainerID: testContainerID,
-			}),
-			usePreview: true,
-		},
-		{
-			name: "AWS EC2",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAWS,
-				conventions.AttributeHostID:        testHostID,
-				conventions.AttributeHostName:      testHostName,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testHostName},
-		},
-		{
-			name: "AWS EC2, preview",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAWS,
-				conventions.AttributeHostID:        testHostID,
-				conventions.AttributeHostName:      testHostName,
-			}),
-			ok:         true,
-			src:        source.Source{Kind: source.HostnameKind, Identifier: testHostID},
-			usePreview: true,
-		},
-		{
-			name: "ECS Fargate",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider:      conventions.AttributeCloudProviderAWS,
-				conventions.AttributeCloudPlatform:      conventions.AttributeCloudPlatformAWSECS,
-				conventions.AttributeAWSECSTaskARN:      "example-task-ARN",
-				conventions.AttributeAWSECSTaskFamily:   "example-task-family",
-				conventions.AttributeAWSECSTaskRevision: "example-task-revision",
-				conventions.AttributeAWSECSLaunchtype:   conventions.AttributeAWSECSLaunchtypeFargate,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.AWSECSFargateKind, Identifier: "example-task-ARN"},
-		},
-		{
-			name: "GCP",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider: conventions.AttributeCloudProviderGCP,
-				conventions.AttributeHostID:        testHostID,
-				conventions.AttributeHostName:      testGCPHostname,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testGCPHostname},
-		},
-		{
-			name: "GCP, preview",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider:  conventions.AttributeCloudProviderGCP,
-				conventions.AttributeHostID:         testHostID,
-				conventions.AttributeHostName:       testGCPHostname,
-				conventions.AttributeCloudAccountID: testCloudAccount,
-			}),
-			usePreview: true,
-			ok:         true,
-			src:        source.Source{Kind: source.HostnameKind, Identifier: testGCPIntegrationHostname},
-		},
-		{
-			name: "azure",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAzure,
-				conventions.AttributeHostID:        testHostID,
-				conventions.AttributeHostName:      testHostName,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testHostName},
-		},
-		{
-			name: "azure, preview",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAzure,
-				conventions.AttributeHostID:        testHostID,
-				conventions.AttributeHostName:      testHostName,
-			}),
-			usePreview: true,
-			ok:         true,
-			src:        source.Source{Kind: source.HostnameKind, Identifier: testHostID},
-		},
-		{
-			name: "host id v. hostname",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				conventions.AttributeHostID:   testHostID,
-				conventions.AttributeHostName: testHostName,
-			}),
-			ok:  true,
-			src: source.Source{Kind: source.HostnameKind, Identifier: testHostID},
-		},
-		{
-			name:  "no hostname",
-			attrs: testutils.NewAttributeMap(map[string]string{}),
-		},
-		{
-			name: "localhost",
-			attrs: testutils.NewAttributeMap(map[string]string{
-				AttributeDatadogHostname: "127.0.0.1",
-			}),
-		},
-	}
-
-	for _, testInstance := range tests {
-		t.Run(testInstance.name, func(t *testing.T) {
-			source, ok := SourceFromAttributes(testInstance.attrs, testInstance.usePreview)
-			assert.Equal(t, testInstance.ok, ok)
-			assert.Equal(t, testInstance.src, source)
-		})
-
-	}
-}
-
 func TestSourceFromAttrs(t *testing.T) {
 	tests := []struct {
 		name  string
@@ -191,6 +47,20 @@ func TestSourceFromAttrs(t *testing.T) {
 		ok  bool
 		src source.Source
 	}{
+		{
+			name: "literal 'host' tag",
+			attrs: testutils.NewAttributeMap(map[string]string{
+				AttributeHost:                       testLiteralHost,
+				AttributeDatadogHostname:            testCustomName,
+				AttributeK8sNodeName:                testNodeName,
+				conventions.AttributeK8SClusterName: testClusterName,
+				conventions.AttributeContainerID:    testContainerID,
+				conventions.AttributeHostID:         testHostID,
+				conventions.AttributeHostName:       testHostName,
+			}),
+			ok:  true,
+			src: source.Source{Kind: source.HostnameKind, Identifier: testLiteralHost},
+		},
 		{
 			name: "custom hostname",
 			attrs: testutils.NewAttributeMap(map[string]string{
@@ -293,6 +163,14 @@ func TestSourceFromAttrs(t *testing.T) {
 	}
 }
 
+func TestLiteralHostNonString(t *testing.T) {
+	attrs := pcommon.NewMap()
+	attrs.PutInt(AttributeHost, 1000)
+	src, ok := SourceFromAttrs(attrs)
+	assert.True(t, ok)
+	assert.Equal(t, source.Source{Kind: source.HostnameKind, Identifier: "1000"}, src)
+}
+
 func TestGetClusterName(t *testing.T) {
 	// OpenTelemetry convention
 	attrs := testutils.NewAttributeMap(map[string]string{
@@ -335,7 +213,7 @@ func TestHostnameKubernetes(t *testing.T) {
 		conventions.AttributeHostID:         testHostID,
 		conventions.AttributeHostName:       testHostName,
 	})
-	hostname, ok := hostnameFromAttributes(attrs, false)
+	hostname, ok := hostnameFromAttributes(attrs)
 	assert.True(t, ok)
 	assert.Equal(t, hostname, "nodeName-clusterName")
 
@@ -346,11 +224,11 @@ func TestHostnameKubernetes(t *testing.T) {
 		conventions.AttributeHostID:      testHostID,
 		conventions.AttributeHostName:    testHostName,
 	})
-	hostname, ok = hostnameFromAttributes(attrs, false)
+	hostname, ok = hostnameFromAttributes(attrs)
 	assert.True(t, ok)
 	assert.Equal(t, hostname, "nodeName")
 
-	// Node name, no cluster name, AWS EC2, no preview
+	// Node name, no cluster name, AWS EC2
 	attrs = testutils.NewAttributeMap(map[string]string{
 		AttributeK8sNodeName:               testNodeName,
 		conventions.AttributeContainerID:   testContainerID,
@@ -358,19 +236,7 @@ func TestHostnameKubernetes(t *testing.T) {
 		conventions.AttributeHostName:      testHostName,
 		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAWS,
 	})
-	hostname, ok = hostnameFromAttributes(attrs, false)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, "nodeName")
-
-	// Node name, no cluster name, AWS EC2, preview
-	attrs = testutils.NewAttributeMap(map[string]string{
-		AttributeK8sNodeName:               testNodeName,
-		conventions.AttributeContainerID:   testContainerID,
-		conventions.AttributeHostID:        testHostID,
-		conventions.AttributeHostName:      testHostName,
-		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAWS,
-	})
-	hostname, ok = hostnameFromAttributes(attrs, true)
+	hostname, ok = hostnameFromAttributes(attrs)
 	assert.True(t, ok)
 	assert.Equal(t, hostname, testHostID)
 
@@ -381,7 +247,7 @@ func TestHostnameKubernetes(t *testing.T) {
 		conventions.AttributeHostID:         testHostID,
 		conventions.AttributeHostName:       testHostName,
 	})
-	hostname, ok = hostnameFromAttributes(attrs, false)
+	hostname, ok = hostnameFromAttributes(attrs)
 	assert.True(t, ok)
 	// cluster name gets ignored, fallback to next option
 	assert.Equal(t, hostname, testHostID)